separate out location types
