+++ core/avs/service/module/src/main/java/org/example/age/module/config/avs/RegisteredSiteConfig.java
@@ -3,7 +3,7 @@ package org.example.age.module.config.avs;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.example.age.data.json.JsonStyle;
 import org.example.age.data.user.AgeThresholds;
-import org.example.age.module.config.common.SiteLocation;
+import org.example.age.module.location.common.SiteLocation;
 import org.immutables.value.Value;
 
 /** Configuration for a site that is registered with the age verification service. */
