+++ core/site/service/endpoint/build.gradle.kts
@@ -39,11 +39,11 @@ dependencies {
     testFixturesImplementation(project(":core:common:service:module"))
     testFixturesApi(project(":core:common:service:endpoint"))
     testFixturesApi(project(":infra:service"))
-    testFixturesApi(testFixtures(project(":module:config:common:test")))
     testFixturesApi(testFixtures(project(":module:config:site:test")))
     testFixturesApi(project(":module:extractor:common:builtin"))
     testFixturesApi(testFixtures(project(":module:extractor:common:test")))
     testFixturesApi(testFixtures(project(":module:key:common:test")))
+    testFixturesApi(testFixtures(project(":module:location:common:test")))
     testFixturesApi(project(":module:store:common:inmemory"))
     testFixturesImplementation("com.fasterxml.jackson.core:jackson-databind")
