+++ core/site/service/endpoint/src/testFixtures/java/org/example/age/service/avs/endpoint/FakeAvsServiceModule.java
@@ -5,10 +5,10 @@ import dagger.Module;
 import io.undertow.server.HttpHandler;
 import org.example.age.api.avs.AvsApi;
 import org.example.age.api.avs.endpoint.AvsApiModule;
-import org.example.age.module.config.common.SiteLocation;
 import org.example.age.module.extractor.common.AccountIdExtractor;
 import org.example.age.module.extractor.common.AuthMatchDataExtractor;
 import org.example.age.module.key.common.RefreshableKeyProvider;
+import org.example.age.module.location.common.SiteLocation;
 import org.example.age.service.avs.verification.internal.FakeAvsVerificationFactoryModule;
 import org.example.age.service.infra.client.RequestDispatcherModule;
