--- module/config/common/test/src/testFixtures/java/org/example/age/module/config/common/test/TestAvsLocationModule.java
@@ -1,21 +1,0 @@
-package org.example.age.module.config.common.test;
-
-import dagger.Module;
-import dagger.Provides;
-import javax.inject.Named;
-import org.example.age.module.config.common.AvsLocation;
-import org.example.age.testing.server.TestServer;
-
-/**
- * Dagger module that publishes a binding for {@link AvsLocation}.
- *
- * <p>Depends on an unbound <code>@Named("avs") {@link TestServer}&lt;?&gt;</code>.</p>
- */
-@Module
-public interface TestAvsLocationModule {
-
-    @Provides
-    static AvsLocation provideAvsLocation(@Named("avs") TestServer<?> avsServer) {
-        return TestLocations.avs(avsServer);
-    }
-}
