+++ module/config/site/test/build.gradle.kts
@@ -13,7 +13,7 @@ dependencies {
     testFixturesApi("com.google.dagger:dagger")
     testFixturesApi("javax.inject:javax.inject")
 
-    testFixturesApi(testFixtures(project(":module:config:common:test")))
+    testFixturesApi(testFixtures(project(":module:location:common:test")))
 
     // test
     testAnnotationProcessor("com.google.dagger:dagger-compiler")
