+++ module/location/common/test/src/testFixtures/java/org/example/age/module/location/common/test/StubSiteLocationModule.java
@@ -1,0 +1,17 @@
+package org.example.age.module.location.common.test;
+
+import dagger.Module;
+import dagger.Provides;
+import javax.inject.Singleton;
+import org.example.age.module.location.common.SiteLocation;
+
+/** Dagger module that publishes a binding for {@link SiteLocation}. */
+@Module
+public interface StubSiteLocationModule {
+
+    @Provides
+    @Singleton
+    static SiteLocation provideSiteLocation() {
+        return TestLocations.stubSite();
+    }
+}
